+++ ThirdPartyAdapters/applovin/applovin/src/main/java/com/applovin/mediation/ApplovinAdapter.java
@@ -1,27 +1,20 @@
 package com.applovin.mediation;
 
-import android.app.Activity;
 import android.content.Context;
 import android.os.Bundle;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
 
 import com.applovin.adview.AppLovinAdView;
-import com.applovin.adview.AppLovinAdViewDisplayErrorCode;
-import com.applovin.adview.AppLovinAdViewEventListener;
 import com.applovin.adview.AppLovinIncentivizedInterstitial;
 import com.applovin.adview.AppLovinInterstitialAd;
 import com.applovin.adview.AppLovinInterstitialAdDialog;
 import com.applovin.sdk.AppLovinAd;
-import com.applovin.sdk.AppLovinAdClickListener;
-import com.applovin.sdk.AppLovinAdDisplayListener;
 import com.applovin.sdk.AppLovinAdLoadListener;
-import com.applovin.sdk.AppLovinAdRewardListener;
 import com.applovin.sdk.AppLovinAdSize;
-import com.applovin.sdk.AppLovinAdVideoPlaybackListener;
-import com.applovin.sdk.AppLovinErrorCodes;
 import com.applovin.sdk.AppLovinSdk;
-import com.applovin.sdk.AppLovinSdkSettings;
+import com.applovin.sdk.AppLovinSdkUtils;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdSize;
 import com.google.android.gms.ads.mediation.MediationAdRequest;
@@ -33,369 +26,343 @@ import com.google.android.gms.ads.mediation.OnContextChangedListener;
 import com.google.android.gms.ads.reward.mediation.MediationRewardedVideoAdAdapter;
 import com.google.android.gms.ads.reward.mediation.MediationRewardedVideoAdListener;
 
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Method;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.LinkedList;
+import java.util.Queue;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import static android.util.Log.DEBUG;
 import static android.util.Log.ERROR;
 
 /**
- * AppLovin SDK interstitial and rewarded video ad adapter for AdMob.
- *
- * @version 7.5.0.0
+ * The {@link ApplovinAdapter} class is used to load AppLovin Banner, interstitial &
+ * rewarded-based video ads and to mediate the callbacks between the AppLovin SDK and the Google
+ * Mobile Ads SDK.
  */
-public class ApplovinAdapter implements MediationBannerAdapter, MediationRewardedVideoAdAdapter,
-        MediationInterstitialAdapter, OnContextChangedListener {
+public class ApplovinAdapter
+        implements MediationBannerAdapter, MediationInterstitialAdapter,
+        MediationRewardedVideoAdAdapter, OnContextChangedListener {
+    private static final boolean LOGGING_ENABLED = true;
+    private static final String DEFAULT_ZONE = "";
 
-    private static final boolean LOGGING_ENABLED = false;
-    private static final String ADAPTER_VERSION = BuildConfig.VERSION_NAME;
-    private static final String SDK_KEY_PARAM = "sdkKey";
-    private static final String PLACEMENT_PARAM = "placement";
+    private static final int BANNER_STANDARD_HEIGHT         = 50;
+    private static final int BANNER_HEIGHT_OFFSET_TOLERANCE = 10;
 
-    private static final AppLovinSdkSettings sSdkSettings = new AppLovinSdkSettings();
+    // Interstitial globals.
+    private static final HashMap<String, Queue<AppLovinAd>> INTERSTITIAL_AD_QUEUES =
+            new HashMap<>();
+    private static final Object INTERSTITIAL_AD_QUEUES_LOCK = new Object();
 
-    private static Map<String, AppLovinAdListener> sPlacementListeners = new HashMap<>();
-    private static List<AppLovinAdListener> sLoadList = new ArrayList<>();
-    private static AppLovinSdk sInterstitialSdk;
-    private static AppLovinAdListener sCurrentlyLoading;
+    // Rewarded video globals.
+    private static final HashMap<String, AppLovinIncentivizedInterstitial> INCENTIVIZED_ADS =
+            new HashMap<>();
+    private static final Object INCENTIVIZED_ADS_LOCK = new Object();
 
-    private boolean mInitialized;
-    private AppLovinSdk mRewardedSdk;
-    private AppLovinAdListener mAppLovinAdListener;
-    private AppLovinIncentivizedInterstitial mRewardedAd;
+    // Parent objects.
+    private AppLovinSdk mSdk;
     private Context mContext;
-    private Bundle mServerParameters;
-    private Bundle mMediationExtras;
-    private MediationRewardedVideoAdListener mRewardedAdMobListener;
-    private MediationInterstitialListener mInterstitialAdMobListener;
+    private Bundle mNetworkExtras;
 
-    AppLovinInterstitialAdDialog mInterstitialAd;
+    // Interstitial objects.
+    private MediationInterstitialListener mMediationInterstitialListener;
 
-    // AppLovin Banner fields
+    // Rewarded Video objects.
+    private final AtomicBoolean mInitialized = new AtomicBoolean();
+    private MediationRewardedVideoAdListener mMediationRewardedVideoAdListener;
+    private AppLovinIncentivizedInterstitial mIncentivizedInterstitial;
+
+    // Banner objects.
     private AppLovinAdView mAdView;
-    private AppLovinSdk mAdViewSdk;
 
-    //region AdMob rewarded video ad mediation methods
-    @Override
-    public void initialize(final Context context,
-                           final MediationAdRequest adRequest,
-                           final String userId,
-                           final MediationRewardedVideoAdListener listener,
-                           final Bundle serverParameters,
-                           final Bundle networkExtras) {
-        if (!validateContext(context)) {
-            log(ERROR, "Unable to request AppLovin rewarded video. Invalid context provided.");
-            listener.onInitializationFailed(this, AdRequest.ERROR_CODE_INVALID_REQUEST);
+    // Controlled fields.
+    private String mPlacement;
+    private String mZoneId;
 
-            return;
-        }
-
-        log(DEBUG, "Initializing AppLovin rewarded video...");
-
-        this.mContext = context;
-        mRewardedAdMobListener = listener;
-        this.mServerParameters = serverParameters;
-
-        if (!mInitialized) {
-            mRewardedSdk = sdkInstance(context);
-
-            mInitialized = true;
-            mRewardedAd = AppLovinIncentivizedInterstitial.create(mRewardedSdk);
-            mAppLovinAdListener = new AppLovinAdListener(this, mRewardedAdMobListener);
-        }
-
-        mRewardedAdMobListener.onInitializationSucceeded(this);
-    }
-
-    @Override
-    public boolean isInitialized() {
-        return mInitialized;
-    }
-
-    @Override
-    public void loadAd(final MediationAdRequest adRequest,
-                       final Bundle serverParameters,
-                       final Bundle networkExtras) {
-        log(DEBUG, "Requesting AppLovin rewarded video");
-
-        if (mRewardedAd.isAdReadyToDisplay()) {
-            mRewardedAdMobListener.onAdLoaded(this);
-        } else {
-            mRewardedAd.preload(mAppLovinAdListener);
-        }
-    }
-
-    @Override
-    public void showVideo() {
-        if (mRewardedAd.isAdReadyToDisplay()) {
-            try {
-                // AppLovin SDK < 7.2.0 uses an Activity, as opposed to Context in >= 7.2.0
-                final Class<?> contextClass =
-                        (AppLovinSdk.VERSION_CODE < 720) ? Activity.class : Context.class;
-                final Method showMethod = AppLovinIncentivizedInterstitial.class.getMethod("show",
-                        contextClass,
-                        String.class,
-                        AppLovinAdRewardListener.class,
-                        AppLovinAdVideoPlaybackListener.class,
-                        AppLovinAdDisplayListener.class,
-                        AppLovinAdClickListener.class);
-
-                try {
-                    if (muteAudio()) {
-                        mRewardedSdk.getSettings().setMuted(muteAudio());
-                    }
-
-                    if (placement() != null) {
-                        showMethod.invoke(mRewardedAd,
-                                this.mContext,
-                                placement(),
-                                mAppLovinAdListener,
-                                mAppLovinAdListener,
-                                mAppLovinAdListener,
-                                mAppLovinAdListener);
-                    } else {
-                        showMethod.invoke(mRewardedAd,
-                                this.mContext,
-                                null,
-                                mAppLovinAdListener,
-                                mAppLovinAdListener,
-                                mAppLovinAdListener,
-                                mAppLovinAdListener);
-                    }
-                } catch (Throwable th) {
-                    log(ERROR, "Unable to invoke show() method from"
-                            + " AppLovinIncentivizedInterstitial.");
-                    mRewardedAdMobListener.onAdOpened(this);
-                    mRewardedAdMobListener.onAdClosed(this);
-                }
-            } catch (Throwable th) {
-                log(ERROR, "Unable to get show() method from AppLovinIncentivizedInterstitial.");
-                mRewardedAdMobListener.onAdOpened(this);
-                mRewardedAdMobListener.onAdClosed(this);
-            }
-        } else {
-            log(ERROR, "Failed to show an AppLovin rewarded video before one was loaded");
-            mRewardedAdMobListener.onAdOpened(this);
-            mRewardedAdMobListener.onAdClosed(this);
-        }
-    }
-    //endregion
-
-    //region Interstitial ad mediation methods
+    //region MediationInterstitialAdapter implementation.
     @Override
     public void requestInterstitialAd(Context context,
-                                      MediationInterstitialListener listener,
+                                      MediationInterstitialListener interstitialListener,
                                       Bundle serverParameters,
                                       MediationAdRequest mediationAdRequest,
-                                      Bundle mediationExtras) {
-        log(DEBUG, "Requesting AppLovin interstitial...");
+                                      Bundle networkExtras) {
+        // Store parent objects.
+        mSdk = AppLovinUtils.retrieveSdk(serverParameters, context);
+        mContext = context;
+        mNetworkExtras = networkExtras;
+        mMediationInterstitialListener = interstitialListener;
 
-        // SDK versions BELOW 7.2.0 require a instance of an Activity to be passed in as the context
-        if (!validateContext(context)) {
-            log(ERROR, "Unable to request AppLovin interstitial. Invalid context provided.");
-            listener.onAdFailedToLoad(this, AdRequest.ERROR_CODE_INVALID_REQUEST);
+        mPlacement = AppLovinUtils.retrievePlacement(serverParameters);
+        mZoneId = AppLovinUtils.retrieveZoneId(serverParameters);
 
-            return;
-        }
+        log(DEBUG, "Requesting interstitial for zone: " + mZoneId + " and placement: "
+                + mPlacement);
 
-        // Store parent objects
-        this.mContext = context;
-        this.mServerParameters = serverParameters;
-        this.mMediationExtras = mediationExtras;
-        mInterstitialAdMobListener = listener;
+        // Create Ad Load listener.
+        final AppLovinAdLoadListener adLoadListener = new AppLovinAdLoadListener() {
+            @Override
+            public void adReceived(final AppLovinAd ad) {
+                log(DEBUG, "Interstitial did load ad: " + ad.getAdIdNumber() + " for zone: "
+                        + mZoneId + " and placement: " + mPlacement);
 
-        if (!sPlacementListeners.containsKey(placement())) {
-            mAppLovinAdListener = new AppLovinAdListener(this, mInterstitialAdMobListener);
-            sPlacementListeners.put(placement(), mAppLovinAdListener);
-        } else {
-            mAppLovinAdListener = sPlacementListeners.get(placement());
+                synchronized (INTERSTITIAL_AD_QUEUES_LOCK) {
+                    Queue<AppLovinAd> preloadedAds = INTERSTITIAL_AD_QUEUES.get(mZoneId);
+                    if (preloadedAds == null) {
+                        preloadedAds = new LinkedList<>();
+                        INTERSTITIAL_AD_QUEUES.put(mZoneId, preloadedAds);
+                    }
 
-            // Update to latest mInterstitialAdMobListener
-            mAppLovinAdListener.updateAdMobListener(mInterstitialAdMobListener);
-        }
+                    preloadedAds.offer(ad);
 
-        sInterstitialSdk = sdkInstance(context);
+                    AppLovinSdkUtils.runOnUiThread(new Runnable() {
+                        @Override
+                        public void run() {
+                            mMediationInterstitialListener.onAdLoaded(ApplovinAdapter.this);
+                        }
+                    });
+                }
+            }
 
-        log(DEBUG, "Adding interstitial request to load list for placement: " + placement());
-        sLoadList.add(mAppLovinAdListener);
-        loadNextAd();
+            @Override
+            public void failedToReceiveAd(final int code) {
+                log(ERROR, "Interstitial failed to load with error: " + code);
 
-        if (!AppLovinAdListener.ADS_QUEUE.isEmpty()) {
-            adLoaded(mAppLovinAdListener);
-            mInterstitialAdMobListener.onAdLoaded(ApplovinAdapter.this);
+                AppLovinSdkUtils.runOnUiThread(new Runnable() {
+                    @Override
+                    public void run() {
+                        mMediationInterstitialListener.onAdFailedToLoad(
+                                ApplovinAdapter.this, AppLovinUtils.toAdMobErrorCode(code));
+                    }
+                });
+            }
+        };
+
+        synchronized (INTERSTITIAL_AD_QUEUES_LOCK) {
+            final Queue<AppLovinAd> queue = INTERSTITIAL_AD_QUEUES.get(mZoneId);
+            if (queue == null || (queue != null && queue.isEmpty())) {
+                // If we don't already have enqueued ads, fetch from SDK.
+
+                if (!TextUtils.isEmpty(mZoneId)) {
+                    mSdk.getAdService().loadNextAdForZoneId(mZoneId, adLoadListener);
+                } else {
+                    mSdk.getAdService().loadNextAd(AppLovinAdSize.INTERSTITIAL, adLoadListener);
+                }
+            } else {
+                log(DEBUG, "Enqueued interstitial found. Finishing load...");
+
+                AppLovinSdkUtils.runOnUiThread(new Runnable() {
+                    @Override
+                    public void run() {
+                        mMediationInterstitialListener.onAdLoaded(ApplovinAdapter.this);
+                    }
+                });
+            }
         }
     }
 
     @Override
     public void showInterstitial() {
-        mAppLovinAdListener = sPlacementListeners.get(placement());
+        synchronized (INTERSTITIAL_AD_QUEUES_LOCK) {
+            // Update mute state.
+            mSdk.getSettings().setMuted(AppLovinUtils.shouldMuteAudio(mNetworkExtras));
 
-        if ((mAppLovinAdListener != null && mAppLovinAdListener.hasAdReady())
-                || AppLovinInterstitialAd.isAdReadyToDisplay(
-                        sInterstitialSdk.getApplicationContext())) {
-            final AppLovinInterstitialAdDialog interstitialAd = createInterstitial();
-            interstitialAd.setAdDisplayListener(mAppLovinAdListener);
-            interstitialAd.setAdClickListener(mAppLovinAdListener);
-            interstitialAd.setAdVideoPlaybackListener(mAppLovinAdListener);
+            final Queue<AppLovinAd> queue = INTERSTITIAL_AD_QUEUES.get(mZoneId);
+            final AppLovinAd dequeuedAd = (queue != null) ? queue.poll() : null;
 
-            // mute audio if that option is set
-            if (muteAudio()) {
-                sInterstitialSdk.getSettings().setMuted(muteAudio());
-            }
+            final AppLovinInterstitialAdDialog interstitialAd =
+                    AppLovinInterstitialAd.create(mSdk, mContext);
 
-            // show with a placement if one is defined
-            if (mAppLovinAdListener != null) {
-                if (mAppLovinAdListener.hasAdReady()) {
-                    interstitialAd.showAndRender(mAppLovinAdListener.dequeueAd(), placement());
-                } else {
-                    interstitialAd.show(placement());
-                }
+            final AppLovinInterstitialAdListener listener =
+                    new AppLovinInterstitialAdListener(this, mMediationInterstitialListener);
+            interstitialAd.setAdDisplayListener(listener);
+            interstitialAd.setAdClickListener(listener);
+            interstitialAd.setAdVideoPlaybackListener(listener);
+
+            if (dequeuedAd != null) {
+                log(DEBUG, "Showing interstitial for zone: " + mZoneId + " placement: "
+                        + mPlacement);
+                interstitialAd.showAndRender(dequeuedAd, mPlacement);
             } else {
-                interstitialAd.show(placement());
+                log(DEBUG, "Attempting to show interstitial before one was loaded");
+
+                // Check if we have a default zone interstitial available.
+                if (TextUtils.isEmpty(mZoneId) && interstitialAd.isAdReadyToDisplay()) {
+                    log(DEBUG, "Showing interstitial preloaded by SDK");
+                    interstitialAd.show(mPlacement);
+                }
+                // TODO: Show ad for zone identifier if exists
+                else {
+                    mMediationInterstitialListener.onAdOpened(this);
+                    mMediationInterstitialListener.onAdClosed(this);
+                }
             }
-        } else {
-            log(ERROR, "Failed to show an AppLovin interstitial before one was loaded."
-                    + " Loading and showing...");
-            mInterstitialAdMobListener.onAdOpened(this);
-            mInterstitialAdMobListener.onAdClosed(this);
         }
-    }
-
-    private AppLovinInterstitialAdDialog createInterstitial() {
-        AppLovinInterstitialAdDialog inter = null;
-
-        try {
-            // AppLovin SDK < 7.2.0 uses an Activity, as opposed to Context in >= 7.2.0
-            final Class<?> contextClass =
-                    (AppLovinSdk.VERSION_CODE < 720) ? Activity.class : Context.class;
-            final Method method =
-                    AppLovinInterstitialAd.class.getMethod(
-                            "create", AppLovinSdk.class, contextClass);
-            inter = (AppLovinInterstitialAdDialog) method.invoke(null, sInterstitialSdk, mContext);
-        } catch (Throwable th) {
-            log(ERROR, "Unable to create AppLovinInterstitialAd.");
-            mInterstitialAdMobListener.onAdFailedToLoad(this, AdRequest.ERROR_CODE_INTERNAL_ERROR);
-        }
-
-        return inter;
     }
     //endregion
 
-
-    //region MediationBannerAdapter implementation.
+    //region MediationRewardedVideoAdAdapter implementation.
     @Override
-    public void requestBannerAd(final Context context,
-                                final MediationBannerListener mediationBannerListener,
-                                final Bundle serverParameters,
-                                final AdSize adSize,
-                                final MediationAdRequest mediationAdRequest,
-                                final Bundle mediationExtras) {
-        // TODO: More elegant solution for retrieving SDK
-        mServerParameters = serverParameters;
-        mAdViewSdk = sdkInstance(context);
+    public void initialize(Context context,
+                           MediationAdRequest adRequest,
+                           String userId,
+                           MediationRewardedVideoAdListener rewardedVideoAdListener,
+                           Bundle serverParameters,
+                           Bundle networkExtras) {
+        log(DEBUG, "Attempting to initialize SDK");
 
-        // SDK versions < 7.1.0 require a instance of an Activity to be passed in as the context
-        if (AppLovinSdk.VERSION_CODE < 710 && !(context instanceof Activity)) {
-            log(ERROR, "Unable to request AppLovin banner. Invalid context provided.");
-            mediationBannerListener.onAdFailedToLoad(this, AdRequest.ERROR_CODE_INVALID_REQUEST);
-
-            return;
+        if (!mInitialized.getAndSet(true)) {
+            // Store parent objects.
+            mSdk = AppLovinUtils.retrieveSdk(serverParameters, context);
+            mContext = context;
+            mNetworkExtras = networkExtras;
+            mMediationRewardedVideoAdListener = rewardedVideoAdListener;
         }
 
-        log(DEBUG, "Requesting AppLovin banner of size: " + adSize);
+        rewardedVideoAdListener.onInitializationSucceeded(this);
+    }
 
-        final AppLovinAdSize appLovinAdSize = appLovinAdSizeFromAdMobAdSize(adSize);
-        if (appLovinAdSize != null) {
-            mAdView = createAdView(appLovinAdSize, context, mediationBannerListener);
+    @Override
+    public boolean isInitialized() {
+        // Please note: AdMob re-uses the adapter for rewarded videos, so be cognizant of states.
+        return mInitialized.get();
+    }
 
-            // Add paranoia null-check
-            if (mAdView != null) {
-                mAdView.setAdLoadListener(new AppLovinAdLoadListener() {
-                    @Override
-                    public void adReceived(final AppLovinAd ad) {
-                        log(DEBUG, "Successfully loaded banner ad");
+    @Override
+    public void loadAd(MediationAdRequest adRequest,
+                       Bundle serverParameters,
+                       Bundle networkExtras) {
+        synchronized (INCENTIVIZED_ADS_LOCK) {
+            mPlacement = AppLovinUtils.retrievePlacement(serverParameters);
+            mZoneId = AppLovinUtils.retrieveZoneId(serverParameters);
 
-                        final String placement = (serverParameters != null)
-                                ? serverParameters.getString(PLACEMENT_PARAM) : null;
-                        mAdView.renderAd(ad, placement);
+            log(DEBUG, "Requesting rewarded video for zone: " + mZoneId + " and placement: "
+                    + mPlacement);
 
-                        mediationBannerListener.onAdLoaded(ApplovinAdapter.this);
-                    }
+            // Check if incentivized ad for zone already exists.
+            if (INCENTIVIZED_ADS.containsKey(mZoneId)) {
+                mIncentivizedInterstitial = INCENTIVIZED_ADS.get(mZoneId);
+            } else {
+                // If this is a default Zone, create the incentivized ad normally
+                if (DEFAULT_ZONE.equals(mZoneId)) {
+                    mIncentivizedInterstitial = AppLovinIncentivizedInterstitial.create(mSdk);
+                }
+                // Otherwise, use the Zones API
+                else {
+                    mIncentivizedInterstitial =
+                            AppLovinIncentivizedInterstitial.create(mZoneId, mSdk);
+                }
 
-                    @Override
-                    public void failedToReceiveAd(final int errorCode) {
-                        log(ERROR, "Failed to load banner ad with code: " + errorCode);
-                        mediationBannerListener.onAdFailedToLoad(
-                                ApplovinAdapter.this, toAdMobErrorCode(errorCode));
-                    }
-                });
-                mAdView.setAdDisplayListener(new AppLovinAdDisplayListener() {
-                    @Override
-                    public void adDisplayed(final AppLovinAd ad) {
-                        log(DEBUG, "Banner displayed");
-                    }
+                INCENTIVIZED_ADS.put(mZoneId, mIncentivizedInterstitial);
+            }
+        }
 
-                    @Override
-                    public void adHidden(final AppLovinAd ad) {
-                        log(DEBUG, "Banner dismissed");
-                    }
-                });
-                mAdView.setAdClickListener(new AppLovinAdClickListener() {
-                    @Override
-                    public void adClicked(final AppLovinAd ad) {
-                        log(DEBUG, "Banner clicked");
+        if (mIncentivizedInterstitial.isAdReadyToDisplay()) {
+            AppLovinSdkUtils.runOnUiThread(new Runnable() {
+                @Override
+                public void run() {
+                    mMediationRewardedVideoAdListener.onAdLoaded(ApplovinAdapter.this);
+                }
+            });
+        } else {
+            // Save placement and zone id, in case the adapter is re-used with new values but the
+            // callbacks are for the old values.
+            final String placement = mPlacement;
+            final String zoneId = mZoneId;
 
-                        mediationBannerListener.onAdClicked(ApplovinAdapter.this);
-                        mediationBannerListener.onAdOpened(ApplovinAdapter.this);
-                        mediationBannerListener.onAdLeftApplication(ApplovinAdapter.this);
-                    }
-                });
+            mIncentivizedInterstitial.preload(new AppLovinAdLoadListener() {
+                @Override
+                public void adReceived(final AppLovinAd ad) {
+                    log(DEBUG, "Rewarded video did load ad: " + ad.getAdIdNumber() + " for zone: "
+                            + zoneId + " and placement: " + placement);
 
-                // As of Android SDK >= 7.3.0, we added a listener for banner events
-                if (AppLovinSdk.VERSION_CODE >= 730) {
-                    mAdView.setAdViewEventListener(new AppLovinAdViewEventListener() {
+                    AppLovinSdkUtils.runOnUiThread(new Runnable() {
                         @Override
-                        public void adOpenedFullscreen(final AppLovinAd ad,
-                                                       final AppLovinAdView adView) {
-                            log(DEBUG, "Banner opened fullscreen");
-                            mediationBannerListener.onAdOpened(ApplovinAdapter.this);
-                        }
-
-                        @Override
-                        public void adClosedFullscreen(final AppLovinAd ad,
-                                                       final AppLovinAdView adView) {
-                            log(DEBUG, "Banner closed fullscreen");
-                            mediationBannerListener.onAdClosed(ApplovinAdapter.this);
-                        }
-
-                        @Override
-                        public void adLeftApplication(final AppLovinAd ad,
-                                                      final AppLovinAdView adView) {
-                            // We will fire onAdLeftApplication() in the adClicked() callback
-                            log(DEBUG, "Banner left application");
-                        }
-
-                        @Override
-                        public void adFailedToDisplay(final AppLovinAd ad,
-                                                      final AppLovinAdView adView,
-                                                      final AppLovinAdViewDisplayErrorCode code) {
-                            log(DEBUG, "Banner failed to display: " + code);
+                        public void run() {
+                            mMediationRewardedVideoAdListener.onAdLoaded(ApplovinAdapter.this);
                         }
                     });
                 }
 
-                mAdView.loadNextAd();
+                @Override
+                public void failedToReceiveAd(final int code) {
+                    log(ERROR, "Rewarded video failed to load with error: " + code);
+
+                    AppLovinSdkUtils.runOnUiThread(new Runnable() {
+                        @Override
+                        public void run() {
+                            mMediationRewardedVideoAdListener.onAdFailedToLoad(ApplovinAdapter.this,
+                                    AppLovinUtils.toAdMobErrorCode(code));
+                        }
+                    });
+                }
+            });
+        }
+    }
+
+    @Override
+    public void showVideo() {
+        if (mIncentivizedInterstitial.isAdReadyToDisplay()) {
+            // Update mute state
+            mSdk.getSettings().setMuted(AppLovinUtils.shouldMuteAudio(mNetworkExtras));
+
+            log(DEBUG, "Showing rewarded video for zone: " + mZoneId + " placement: " + mPlacement);
+
+            final AppLovinIncentivizedAdListener listener =
+                    new AppLovinIncentivizedAdListener(this, mMediationRewardedVideoAdListener);
+            mIncentivizedInterstitial.show(mContext, mPlacement, listener, listener, listener,
+                    listener);
+        } else {
+            log(DEBUG, "Attempting to show rewarded video before one was loaded");
+
+            // TODO: Add support for checking default SDK-preloaded ad
+            mMediationRewardedVideoAdListener.onAdOpened(this);
+            mMediationRewardedVideoAdListener.onAdClosed(this);
+        }
+    }
+    //endregion
+
+    //region MediationBannerAdapter implementation.
+    @Override
+    public void requestBannerAd(Context context,
+                                final MediationBannerListener mediationBannerListener,
+                                Bundle serverParameters,
+                                AdSize adSize,
+                                MediationAdRequest mediationAdRequest,
+                                Bundle networkExtras) {
+        // Store parent objects
+        mSdk = AppLovinUtils.retrieveSdk(serverParameters, context);
+
+        mPlacement = AppLovinUtils.retrievePlacement(serverParameters);
+        mZoneId = AppLovinUtils.retrieveZoneId(serverParameters);
+
+        log(DEBUG, "Requesting banner of size " + adSize + " for zone: "
+                + mZoneId + " and placement: " + mPlacement);
+
+        // Convert requested size to AppLovin Ad Size.
+        final AppLovinAdSize appLovinAdSize = appLovinAdSizeFromAdMobAdSize(adSize);
+        if (appLovinAdSize != null) {
+            mAdView = new AppLovinAdView(mSdk, appLovinAdSize, context);
+
+            final AppLovinBannerAdListener listener = new AppLovinBannerAdListener(
+                    mZoneId, mPlacement, mAdView, this, mediationBannerListener);
+            mAdView.setAdDisplayListener(listener);
+            mAdView.setAdClickListener(listener);
+            mAdView.setAdViewEventListener(listener);
+
+            if (!TextUtils.isEmpty(mZoneId)) {
+                mSdk.getAdService().loadNextAdForZoneId(mZoneId, listener);
             } else {
-                log(ERROR, "Unable to request AppLovin banner");
-                mediationBannerListener.onAdFailedToLoad(this, AdRequest.ERROR_CODE_INTERNAL_ERROR);
+                mSdk.getAdService().loadNextAd(appLovinAdSize, listener);
             }
         } else {
-            log(ERROR, "Unable to request AppLovin banner");
-            mediationBannerListener.onAdFailedToLoad(this, AdRequest.ERROR_CODE_INVALID_REQUEST);
+            log(ERROR, "Failed to request banner with unsupported size");
+
+            AppLovinSdkUtils.runOnUiThread(new Runnable() {
+                @Override
+                public void run() {
+                    mediationBannerListener.onAdFailedToLoad(
+                            ApplovinAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
+                }
+            });
         }
     }
 
@@ -405,9 +372,11 @@ public class ApplovinAdapter implements MediationBannerAdapter, MediationRewarde
     }
     //endregion
 
-    //region Banner utility methods
-    private AppLovinAdSize appLovinAdSizeFromAdMobAdSize(final AdSize adSize) {
-        if (AdSize.BANNER.equals(adSize)) {
+    private AppLovinAdSize appLovinAdSizeFromAdMobAdSize(AdSize adSize) {
+        final boolean isSmartBanner = (adSize.getWidth() == AdSize.FULL_WIDTH) &&
+                (adSize.getHeight() == AdSize.AUTO_HEIGHT);
+
+        if (AdSize.BANNER.equals(adSize) || AdSize.LARGE_BANNER.equals(adSize) || isSmartBanner) {
             return AppLovinAdSize.BANNER;
         } else if (AdSize.MEDIUM_RECTANGLE.equals(adSize)) {
             return AppLovinAdSize.MREC;
@@ -415,33 +384,16 @@ public class ApplovinAdapter implements MediationBannerAdapter, MediationRewarde
             return AppLovinAdSize.LEADER;
         }
 
+        // Assume fluid width, and check for height with offset tolerance
+        final int offset = Math.abs( BANNER_STANDARD_HEIGHT - adSize.getHeight() );
+        if (offset <= BANNER_HEIGHT_OFFSET_TOLERANCE)  {
+            return AppLovinAdSize.BANNER;
+        }
+
         return null;
     }
 
-    private AppLovinAdView createAdView(final AppLovinAdSize size,
-                                        final Context parentContext,
-                                        final MediationBannerListener customEventBannerListener) {
-        AppLovinAdView adView = null;
-
-        try {
-            // AppLovin SDK < 7.1.0 uses an Activity, as opposed to Context in >= 7.1.0
-            final Class<?> contextClass =
-                    (AppLovinSdk.VERSION_CODE < 710) ? Activity.class : Context.class;
-            final Constructor<?> constructor = AppLovinAdView.class
-                    .getConstructor(AppLovinSdk.class, AppLovinAdSize.class, contextClass);
-
-            adView = (AppLovinAdView) constructor.newInstance(mAdViewSdk, size, parentContext);
-        } catch (Throwable th) {
-            log(ERROR, "Unable to get create AppLovinAdView.");
-            customEventBannerListener.onAdFailedToLoad(this,
-                    AdRequest.ERROR_CODE_INVALID_REQUEST);
-        }
-
-        return adView;
-    }
-    //endregion
-
-    //region MediationAdapter methods.
+    //region MediationAdapter.
     @Override
     public void onPause() {
     }
@@ -455,92 +407,19 @@ public class ApplovinAdapter implements MediationBannerAdapter, MediationRewarde
     }
     //endregion
 
+    // OnContextChangedListener Method.
     @Override
-    public void onContextChanged(final Context context) {
+    public void onContextChanged(Context context) {
         if (context != null) {
             log(DEBUG, "Context changed: " + context);
-            this.mContext = context;
+            mContext = context;
         }
     }
 
-    static synchronized void adLoaded(AppLovinAdListener appLovinAdListener) {
-        sLoadList.remove(appLovinAdListener);
-        sCurrentlyLoading = null;
-        loadNextAd();
-    }
-
-    static synchronized void adLoadFailed(AppLovinAdListener appLovinAdListener) {
-        sLoadList.remove(appLovinAdListener);
-        sCurrentlyLoading = null;
-        loadNextAd();
-    }
-
-    private static synchronized void loadNextAd() {
-        if (sLoadList.size() > 0 && ((sCurrentlyLoading != null
-                && sCurrentlyLoading != sLoadList.get(0)) || sCurrentlyLoading == null)) {
-            sCurrentlyLoading = sLoadList.get(0);
-            log(DEBUG, "Loading next interstitial ad from load list");
-            sInterstitialSdk.getAdService().loadNextAd(
-                    AppLovinAdSize.INTERSTITIAL, sLoadList.get(0));
-        } else if (sLoadList.size() == 0) {
-            log(DEBUG, "No more interstitials to load");
-        } else {
-            log(DEBUG, "Another ad is currently loading");
-        }
-    }
-
-    private String placement() {
-        return mServerParameters.getString(PLACEMENT_PARAM);
-    }
-
-    private boolean muteAudio() {
-        if (mMediationExtras != null) {
-            return mMediationExtras.getBoolean(AppLovinExtrasBundleBuilder.MUTE_AUDIO, false);
-        }
-
-        return false;
-    }
-
-    private AppLovinSdk sdkInstance(final Context context) {
-        String sdkKey = mServerParameters.getString(SDK_KEY_PARAM);
-        AppLovinSdk sdk;
-
-        if (sdkKey != null && !sdkKey.equals("")) {
-            sdk = AppLovinSdk.getInstance(sdkKey, sSdkSettings, context);
-        } else {
-            sdk = AppLovinSdk.getInstance(context);
-        }
-
-        sdk.setPluginVersion(ADAPTER_VERSION);
-
-        return sdk;
-    }
-
-    //region Utility Methods
-    private static boolean validateContext(Context context) {
-        // SDK versions BELOW 7.2.0 require a instance of an Activity to be passed in as the context
-        if (AppLovinSdk.VERSION_CODE < 720 && !(context instanceof Activity)) {
-            return false;
-        }
-
-        return true;
-    }
-
-    static void log(final int priority, final String message) {
+    // Logging
+    static void log(int priority, final String message) {
         if (LOGGING_ENABLED) {
             Log.println(priority, "AppLovinAdapter", message);
         }
     }
-
-    static int toAdMobErrorCode(final int applovinErrorCode) {
-        if (applovinErrorCode == AppLovinErrorCodes.NO_FILL) {
-            return AdRequest.ERROR_CODE_NO_FILL;
-        } else if (applovinErrorCode == AppLovinErrorCodes.NO_NETWORK
-                || applovinErrorCode == AppLovinErrorCodes.FETCH_AD_TIMEOUT) {
-            return AdRequest.ERROR_CODE_NETWORK_ERROR;
-        } else {
-            return AdRequest.ERROR_CODE_INTERNAL_ERROR;
-        }
-    }
-    //endregion
 }