+++ ThirdPartyAdapters/facebook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Facebook Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 4.28.0.0
+- Verified compatibility with Facebook SDK v4.28.0.
+
+## 4.27.1.0
+- Verified compatibility with Facebook SDK v4.27.1.
+
 ## 4.27.0.0
 - Verified compatibility with Facebook SDK v4.27.0.