+++ ThirdPartyAdapters/ironsource/CHANGELOG.md
@@ -1,4 +1,8 @@
 # ironSource Adapter for Google Mobile Ads SDK for Android Changelog
 
+## Version 6.7.7.0
+- Initiate IronSource SDK once per adunit
+- Add option to pass activity to IronSource SDK in 'onPause'\'onResume' methods
+
 ## Version 6.7.5.0
 - Initial release!